MODORG-75: Misleading warning "checkOperationsRestrictions:: unitIds is empty"
https://folio-org.atlassian.net/browse/MODORG-75

Access to the organization record can be restricted a list of AcqUnits so that a user must be member of at least one of these units to get access.

If an organization record is not restricted the list of AcqUnits is empty.

Actual:

mod-organization writes this WARN message into the mod-organization log:
checkOperationsRestrictions:: unitIds is empty

Expected:

That message should be a DEBUG message because unrestricted organizations are common.

Solution:

Change log level from WARN to DEBUG.

Convert the else branch of the if-else clause into a guard clause (separate if clause) for readability.